pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chimera_llama_grpc"
description = "chimera_llama_grpc"
keywords = ["chimera_llama_grpc"]
requires-python = ">=3.8"
dependencies = [
    "chimera_llm_proto",
    "click",
    "anyio",
    "pydantic>=2",
    "loguru",
    "importlib_metadata",
    # LLAMA
    "torch",
    "fairscale",
    "fire",
    "sentencepiece",
]
dynamic = ["version"]
classifiers = [
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
]
[project.optional-dependencies]
test = ["pytest", "pytest-grpc"]

[project.scripts]
chimera-llama-grpc = "chimera_llama_grpc.cli:cli"

[[project.authors]]
name = "wh1isper"
email = "9573586@qq.com"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.license]
text = "BSD 3-Clause License"

[project.urls]
Source = "https://github.com/wh1isper/chimera_llama_grpc"

[tool.check-manifest]
ignore = [".*"]

[tool.hatch.version]
path = "chimera_llama_grpc/__init__.py"
